// hackathon-client/src/app/models/save-prescription.model.ts
export interface SavePrescription {
  symptoms: string;
  drName: string;
  ptName: string;
  daignosis: string;
  advice: string;
  medicines: string;
  temperature: string;
  respiratoryRate: string;
  bloodPressureMin: string;
  bloodPressureMax: string;
  heartRate: string;
  uniqueId: number;
}